fix(nodes): polish the code placement for node info sync
